Move Mini Controller definitions to Classic files
These are closely related, so no reason to keep them separated - especially with such little code.
