+++ src/controllers/ClassicController.cpp
@@ -181,4 +181,85 @@ void ClassicController_Shared::printDebug(Print& output) const {
 	output.println(buffer);
 }
 
+// ######### Mini Controller Classes #########
+
+boolean NESMiniController_Shared::isKnockoff() const {
+	// The NES knockoffs I've come across seem to display the same unchanging pattern
+	// for the first six control bytes:
+	//
+	//     0x81, 0x81, 0x81, 0x81, 0x00, 0x00
+	//
+	// This is mostly garbage data that doesn't line up with the Classic Controller
+	// at all. Using the library's debug output, here is what it translates to:
+	//
+	//    <^v> | -H+ | ABXY L : (1, 1) R : (21, 1) | LT : 4X RT : 1X Z : LR
+	//
+	// You'll notice a few things. ALL possible buttons are pressed. The left analog 
+	// stick is completely down and to the left, while the right analog stick is down
+	// and to the right. On the back, both trigger "fully depressed" buttons are
+	// down, and yet the analog trigger values are very low. In short, this is a
+	// difficult if not impossible state for a normal Classic Controller to be in.
+	// Because of that, we can reasonably assume that if the bytes match this then the
+	// connected controller is an NES Knockoff, and can be treated accordingly. 
+
+	return getControlData(0) == 0x81 &&  // RX 4:3, LX
+	       getControlData(1) == 0x81 &&  // RX 2:1, LY
+	       getControlData(2) == 0x81 &&  // RX 0, LT 4:3, RY
+	       getControlData(3) == 0x81 &&  // LT 2:0, RT
+	       getControlData(4) == 0x00 &&  // Button packet 1 (all pressed)
+	       getControlData(5) == 0x00;    // Button packet 2 (all pressed)
+}
+
+void NESMiniController_Shared::printDebug(Print& output) const {
+	const char fillCharacter = '_';
+
+	output.print("NES ");
+
+	output.print(dpadLeft() ? '<' : fillCharacter);
+	output.print(dpadUp() ? '^' : fillCharacter);
+	output.print(dpadDown() ? 'v' : fillCharacter);
+	output.print(dpadRight() ? '>' : fillCharacter);
+	output.print(" | ");
+
+	buttonSelect() ? (void)output.print("SEL") : NintendoExtensionCtrl::printRepeat(fillCharacter, 3, output);
+	output.print(' ');
+
+	buttonStart() ? (void)output.print("STR") : NintendoExtensionCtrl::printRepeat(fillCharacter, 3, output);
+	output.print(" | ");
+
+	output.print(buttonB() ? 'B' : fillCharacter);
+	output.print(buttonA() ? 'A' : fillCharacter);
+
+	output.println();
+}
+
+void SNESMiniController_Shared::printDebug(Print& output) const {
+	const char fillCharacter = '_';
+
+	output.print("SNES ");
+
+	output.print(dpadLeft() ? '<' : fillCharacter);
+	output.print(dpadUp() ? '^' : fillCharacter);
+	output.print(dpadDown() ? 'v' : fillCharacter);
+	output.print(dpadRight() ? '>' : fillCharacter);
+	output.print(" | ");
+
+	buttonSelect() ? (void)output.print("SEL") : NintendoExtensionCtrl::printRepeat(fillCharacter, 3, output);
+	output.print(' ');
+
+	buttonStart() ? (void)output.print("STR") : NintendoExtensionCtrl::printRepeat(fillCharacter, 3, output);
+	output.print(" | ");
+
+	output.print(buttonA() ? 'A' : fillCharacter);
+	output.print(buttonB() ? 'B' : fillCharacter);
+	output.print(buttonX() ? 'X' : fillCharacter);
+	output.print(buttonY() ? 'Y' : fillCharacter);
+	output.print(" | ");
+
+	output.print(buttonL() ? 'L' : fillCharacter);
+	output.print(buttonR() ? 'R' : fillCharacter);
+
+	output.println();
+}
+
 }  // End "NintendoExtensionCtrl" namespace
