--- src/controllers/SNESMiniController.h
@@ -1,33 +1,0 @@
-/*
-*  Project     Nintendo Extension Controller Library
-*  @author     David Madison
-*  @link       github.com/dmadison/NintendoExtensionCtrl
-*  @license    LGPLv3 - Copyright (c) 2018 David Madison
-*
-*  This file is part of the Nintendo Extension Controller Library.
-*
-*  This program is free software: you can redistribute it and/or modify
-*  it under the terms of the GNU Lesser General Public License as published by
-*  the Free Software Foundation, either version 3 of the License, or
-*  (at your option) any later version.
-*
-*  This program is distributed in the hope that it will be useful,
-*  but WITHOUT ANY WARRANTY; without even the implied warranty of
-*  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*  GNU Lesser General Public License for more details.
-*
-*  You should have received a copy of the GNU Lesser General Public License
-*  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef NXC_SNESMiniController_h
-#define NXC_SNESMiniController_h
-
-#include "ClassicController.h"
-
-class SNESMiniController : public ClassicController {
-public:
-	void printDebug(Print& output = NXC_SERIAL_DEFAULT) const;
-};
-
-#endif
